updater: support flashing from notifications
merge downloads table into the others, downloads are
now tracked via extra columns. this has the added benefit
of not nuking the table on each update check, retaining
any downloaded files in the list even if they have been
removed from the server

a side effect of this is new entries are added to the bottom of
the table thus appeared at the bottom of the listview (out of order).
to fix this i implemented api v2 which sends us the manifest dated
oldest to newest (reverse of v1) so new entries added to the
end of the table are now in the right order and to sort properly
(newest firest) we simply loop backwards.

most of backend was refactored to pass the ManifestEntry around
as argument, this was required for flash activity but makes
sense to implement allround.

fix the update manifest error message showing 4 times

proper expand / collapse drawables

calculate local md5sum on download complete and store in db
to prevent running everytime we view the file

Change-Id: Ib0671faa0a0a67e283aa76664621a8b5f8d33aeb